added useSwipe
