+++ src/modules/ProductDetailsPage/Galery/Gallery.tsx
@@ -1,36 +1,15 @@
 import classNames from 'classnames';
 import style from './Gallery.module.scss';
-import React, { useRef, useState } from 'react';
+import React from 'react';
+import { useSwipe } from '../../../hook/useSwipe';
 
 type Props = {
   images: string[];
 };
 
 export const Gallery: React.FC<Props> = ({ images }) => {
-  const [currentIndex, setCurrentIndex] = useState(0);
-  const startXRef = useRef(0);
-
-  const nextBunner = () =>
-    setCurrentIndex(prevIndex => (prevIndex + 1) % images.length);
-
-  const previosBunner = () =>
-    setCurrentIndex(
-      prevIndex => (prevIndex - 1 + images.length) % images.length,
-    );
-
-  const handleTouchStart = (event: React.TouchEvent) => {
-    startXRef.current = event.touches[0].clientX;
-  };
-
-  const handleTouchEnd = (event: React.TouchEvent) => {
-    const endX = event.changedTouches[0].clientX;
-
-    if (startXRef.current - endX > 50) {
-      nextBunner();
-    } else if (endX - startXRef.current > 50) {
-      previosBunner();
-    }
-  };
+  const { handleTouchStart, handleTouchEnd, currentIndex, setCurrentIndex } =
+    useSwipe({ images });
 
   return (
     <div className={style.gallery}>
